train-search-bot-taiwan/requirements.txt

filelock==3.0.8
Flask==1.0.2
Flask-Cors==3.0.6
gevent==1.3.6
Werkzeug==0.11.15
line-bot-sdk==1.8.0
gunicorn==19.9.0
pytz==2013b
